Language.Contract.Check

Description

Documentation

type MonadTypeCheck m = (MonadReader ([Type], [Term]) m, MonadFail m, MonadIO m)

Type checker monad.

runTypeCheck :: MaybeT (ReaderT ([Type], [Term]) IO) a -> IO (Maybe a)

Run the type checker monad, get a Maybe result.

assert :: MonadFail m => Bool -> m ()

Assertion, fail with MonadFail. guard from Prelude uses Alternative, so not suitable here.

prove :: MonadTypeCheck m => Term -> m ()

Prove a term in the type checker monad.

typeOf :: MonadTypeCheck m => Term -> m Type

The type of a term, within a specific context.

eval1 :: [Term] -> Term -> Term

One-step evaluation of a term.

evalWith :: [Term] -> Term -> Term

Full evaluation of a term.

eval :: Term -> Term

Full evaluation of a term.