Language.Contract.Parser

Description

Documentation

parseTerm :: String -> Either ParseError Term

Parse a term, for parsing many terms see parseFile.

parseFile :: String -> String -> Either ParseError [(String, Term)]

Parse a file, bindings in the form let name = term;.